refactor(compose): extract core plugin logic to enable addition of compose capabilities
for #2
